refactor(keystore): one file per keystore under core/keystores
The contract stays in core/Keystore.ts; implementations move to
core/keystores/, mirroring core/adapters/. The hexKeystore helper is
inlined into p256 and secp256k1 — duplicating trivial code beats a
shared indirection for two call sites. Docs gain a page per keystore,
mirroring the Storage reference layout.
