+++ site/src/pages/docs/api/keystore.mdx
@@ -15,14 +15,24 @@ Keystores hold key material; [`Storage`](/docs/api/storage) persists provider st
 
 <Cards>
   <Card
+    description="Pure-JS P-256 keys"
+    icon="lucide:file-key"
+    to="/docs/api/keystore.p256"
+    title="Keystore.p256"
+  />
+  <Card
+    description="Pure-JS secp256k1 keys"
+    icon="lucide:file-key"
+    to="/docs/api/keystore.secp256k1"
+    title="Keystore.secp256k1"
+  />
+  <Card
     description="WebCrypto P-256 keys — the built-in default"
     icon="lucide:key-round"
     to="/docs/api/keystore.webCryptoP256"
     title="Keystore.webCryptoP256"
   />
 </Cards>
-
-`Keystore.p256()` and `Keystore.secp256k1()` are pure-JS keystores: the private key is held as hex in the handle, so at-rest protection is the storage adapter and signing runs in JavaScript. Prefer `webCryptoP256` or a hardware keystore where one is available.
 
 ## Interface
