+++ site/src/pages/docs/api/keystore.p256.mdx
@@ -1,0 +1,22 @@
+---
+title: Keystore.p256
+description: Pure-JS P-256 keystore.
+---
+
+# `Keystore.p256`
+
+Creates a pure-JS P-256 [`Keystore`](/docs/api/keystore). The private key is held as hex in the handle, so at-rest protection is the storage adapter and signing runs in JavaScript.
+
+Prefer [`Keystore.webCryptoP256`](/docs/api/keystore.webCryptoP256) or a hardware keystore where a WebCrypto implementation is available; this exists chiefly as an adapter default for environments without one.
+
+## Usage
+
+```ts twoslash
+import { Keystore, Provider } from 'accounts'
+
+const provider = Provider.create({
+  accessKey: {
+    keystores: { p256: Keystore.p256() }, // [!code focus]
+  },
+})
+```
