+++ site/src/pages/docs/api/keystore.secp256k1.mdx
@@ -1,0 +1,22 @@
+---
+title: Keystore.secp256k1
+description: Pure-JS secp256k1 keystore.
+---
+
+# `Keystore.secp256k1`
+
+Creates a pure-JS secp256k1 [`Keystore`](/docs/api/keystore). The private key is held as hex in the handle, so at-rest protection is the storage adapter and signing runs in JavaScript.
+
+secp256k1 signatures are the chain's cheapest envelope, making this the economical default for environments without a WebCrypto implementation — the React Native and CLI adapters supply it as their default keystore. When `authorizeAccessKey` does not request a key type, `secp256k1` is preferred when a keystore for it is configured.
+
+## Usage
+
+```ts twoslash
+import { Keystore, Provider } from 'accounts'
+
+const provider = Provider.create({
+  accessKey: {
+    keystores: { secp256k1: Keystore.secp256k1() }, // [!code focus]
+  },
+})
+```
