+++ site/vocs.config.ts
@@ -204,6 +204,8 @@ const config: Config = defineConfig({
             collapsed: true,
             items: [
               { text: 'Overview', link: '/docs/api/keystore' },
+              { text: '.p256', link: '/docs/api/keystore.p256' },
+              { text: '.secp256k1', link: '/docs/api/keystore.secp256k1' },
               { text: '.webCryptoP256', link: '/docs/api/keystore.webCryptoP256' },
             ],
           },
