+++ src/core/Provider.ts
@@ -1,14 +1,7 @@
 import { announceProvider } from 'mipd'
 import { Challenge } from 'mppx'
 import { Mppx, tempo as mppx_tempo } from 'mppx/client'
-import {
-  Address,
-  Hash,
-  Hex,
-  Json,
-  Provider as ox_Provider,
-  RpcResponse,
-} from 'ox'
+import { Address, Hash, Hex, Json, Provider as ox_Provider, RpcResponse } from 'ox'
 import { KeyAuthorization } from 'ox/tempo'
 import {
   createWalletClient,
